requirements.txt
streamlit
pandas
langchain-groq
langchain-experimental
matplotlib
seaborn
python-dotenv